test: 단일 API 테스트로 분리
